Created a new Rails environment called "ci". Travis-CI builds run in this environment.
It has the same configuration as the "test" environment except for the Log4r config: in the "ci" environment Log4r only outputs to stdout.

This intends to solve the problem we were having in which Travis-CI builds failed since migrating to Log4r, because the "log" directory is not writable.