Require 'factory_girl_rails' instead of 'factory_girl' in spec_helper.
Hopefully this fixes a Travis-CI error in which some factories are not defined.

As a side note, I just don't see why my own build environment is so different from the one Travis runs build in. I'm beginning to think going from a private Jenkins installation to Travis may not have been a good idea.