Changed default Devise paths.
URLs are generally friendlier. For example the sign_in URL instead of /users/sign_in now is /login; the edit registration URL is /profile instead of /users/registration/edit, etc.

Also all forms send data (via POST, PUT or DELETE) to the same URL we have to GET to get the form. This means that if a form is submitted with errors the user stays in the same URL, and can reload the page without errors. With the default Devise routes, when a user submits a form with errors he's sent to a different URL, which means that he cannot reload (or use the language switch button, which reloads the page changing the locale param) without getting ugly routing errors. It's quite surprising that Devise ships with this default buggy behavior.

It seems some of the Devise URL helpers stop working after this change. I've opted for using Rails route helpers directly instead of the Devise helpers. It seems Devise helpers are mainly useful for apps that have several types of authenticable resources (for instance Users, Admins), and I'm only using a single Users authenticable resource anyway.

For more about customizing Devise paths see:

	https://github.com/plataformatec/devise/wiki/How-To:-Customize-the-redirect-after-a-user-edits-their-profile