Fixed broken tests.
There were two problems:

- FactoryGirl was generating URLs with underscores for the feeds. Open-uri throws an error when asked to open such an URL.

- The acceptance test for the feeds view failed because it was trying to open an URL generated by FactoryGirl which, of course, doesn't actually exist (nor does it have a RSS document). For the moment I've changed the test so that real RSS URLs are used; this means that these URLs are actually retrieved during acceptance testing, which is not something I'm happy with. I will change it as soon as possible: tests should never depend on external services or APIs.

The possible long-term solution for the second problem could be to use the RestClient class instead of Open-uri, and substitute it with a mock during acceptance testing.

Or better yet, maybe we should retrieve RSS feeds with a background job instead of during a user process; this way acceptance testing would never trigger an HTTO download. A side benefit would be better performance for the user processes, at the cost of increased complexity in the code.