Added dependency on the Feedzirra gem.
It seems latest stable version (0.1.3) is not compatible with latest stable Rails (3.2.13); they have conflicting dependency versions. I'm fetching Feedzirra v.0.2.0.rc2 from github, it was tagged almost a year ago. This version is compatible with latest stable Rails.

I'm forced to look for a feed parsing gem because the Atom parser included in Ruby stdlib throws an errow when parsing an Atom feed without an author node (like for instance the extremely popular XKCD atom feed at http://www.xkcd.com/atom.xml). From the specification it seems that element is mandatory, but I cannot afford to be incompatible with a popular feed just because it misses that tag.