Added Entry model, along with its migration and FactoryGirl factory. Changed Feed model to reflect that entries always belong to a feed.
When fetching and updating feeds, we will save entries in the database. Each time a user goes to a feed, the feed will not be actually fetched, but rather the entries saved in the DB will be shown. Feed updating will happen in the background, using a background job queue (Resque).

This has many advantages:

- Better performance, the app does not have to fetch/parse an XML to show feed entries to the use, it only has to read them from the database.

- In the future the app will store information about read/unread entries for each user.

- Entries have to be fetched only once and are then available for all users suscribed to the feed.