Added field "fetch_url" to Feed model.
The field "url" will be the URL (present in the feed XML) to which the user will be linked when reading the feed. Normally it's the URL of the website that originates the feed.

The field "fetch_url" will be the URL from which the XML of the feed is fetched.