Show an error when the Ajax call to refresh a feed fails for any reason.
It's been necessary to change the notice and alert divs. Instead of closing them when clicking on the X button with the Bootstrap javascript (which totally removes from the DOM the HTML of the alert), now we bind an event manager that adds the "hidden" class to the div. This way the alert and notice divs are always present and can be manipulated (made visible or invisible, modified their text etc) from javascript.