No feed is loaded at startup, until the user clicks on one of the menu options in the sidebar.
When first loading the feeds view, some kind of landing page will be shown (for the moment it's just a placeholder text). No feed entries nor feed title are shown, and the buttons (refresh etc) are disabled. Only when the user clicks on one of the feeds in the sidebar the title will be shown, entries will be loaded (via Ajax) and buttons will be enabled.