Override the default "accordion-group" CSS class style.
This is a Bootstrap class that is necessary for the entries list to act as an accordion (only one entry can be open at a given time). However it also adds an unnecessary border that mucks up our design. We keep the class, otherwise the accordion javascript doesn't work, but we override that border.