User can now subscribe to a feed by submitting the URL of a webpage which has an Atom feed autodiscovery enabled.
Currently the href attribute of the <link rel="alternate"> tag that enables feed autodiscovery must contain a full URL (with FQDN). Support for the case in which the href attribute contains only a subpath (without FQDN) will be added in the future.

Not yet implemented: autodiscovery of webpages with generic <link rel="feed"> tag.