Added support for feed autodiscovery when the link to the feed has a path instead of a full URL in the href; for example, "/feed.php" instead of "http://webpage.com/feed.php".
In this case the uri-scheme (http, https) and fqdn of the host (e.g. "webpage.com") are assumed to be the same as those of the webpage. A full URL is saved in the fetch_url field of the feed object; in our example, "http://webpage.com/feed.php".

This is necessary because when the feed is fetched again in the future the app needs to know the full URL, a path in this case is not enough to fetch the feed.