Much better organization and DRYing of the javascript timers that autohide notice divs.
Instead of separate event handlers, a single function is called to show+autohide after 5 seconds a notice/alert. Thanks to the magic of closures, autoclose timers run independently.

It also works correctly when the same autoclose timer is invoked repeatedly. E.g. when a user clicks on a feed without entries a notice is shown; if he immediately clicks on another feed without entries the notice will stay onscreen but the timer will be reset to 5 seconds (actually a new timer will be started and the old cleared, but for the user the experience is a timer reset).