Alert/notice divs appear at the bottom of the page.
They have position "fixed" in CSS, which means they stay at the bottom even when scrolling.

Previously they were at the top with static positioning, which meant that the interface "jumped up" when a notice closed. This was quite annoying. Hopefully the new position is less distracting while at the same time clear enough for the user.

The width and offset of the AJAX notices has been chosen so that they overlap with the entries list but are a bit wider on the left side. This helps draw the eye to them and makes them easier to read.