User can unsubscribe from a feed by clicking on the Unsubscribe button and accepting the confirmation popup.
Some acceptance tests are currently failing because of a change in the generated HTML (the <li> for each feed entry in the sidebar now does not have an id generated from the feed id, because it lead to repeated ids if the feed was in a folder; instead data-* attributes are used). The failing tests relied on the old HTML ids to check for the presence of some elements in the page; next commit will fix them.

Some things are still unimplemented: showing alerts if there's a problem, disabling the Unsubscribing button when reading a whole folder, deleting the feed if no users are subscribed to it anymore, showing the main "welcome&statistics" page after deleting.

This last one is interesting. Currently the welcome&statistics view is unimplemented. We're currently making the user read the whole list of entries for all feeds after unsubscribing from a feed, but this is not optimal because this list of entries is the longest one to load. It could get annoying for the user!