WIP: add a button that enables the user to associate a subscribed feed with a folder (already existing or created on the spot) or remove it from all folders.
Added callbacks to Feed and Folder methods to ensure that a given user can only associate a feed with a single folder. This means that if the user wants to associate a feed with a different folder, it will first have to be  removed from its current folder. This is a behavior that may change in the future depending on user feedback, but for the moment it seems the simplest behavior.

Implemented: The button is a dropdown that shows the list of folders that belong to the user, as well as "No folder" and "New folder" entries. A tick icon will be visible besides the folder to which the current feed belongs, or besides the "No folder" entry if the feed is not associated with any folders.

Not yet implemented: Associating a feed with an already existing folder, associating a feed with a new folder (created on the spot), removing a feed from all folders (by clicking on the No Folder link), deleting a folder which has no more feeds associated.

All user actions only affect the association of feeds subscribed by the current user to folders owned by the current user. A user cannot affect folders owned by other users at all.