Added routing and methods in Folder and FolderController that allow a user to associate a feed to a folder.
The user must be subscribed to the feed and the folder must be owned by the user. If the feed is already in another folder that belongs to the same user, it is removed from that folder before associating with the new folder. This means that from the point of view of a given user, a feed can only be associated with a single folder. This behaviour has been chosen for simplicity, but may change in the future depending on user feedback.