Fixed refreshing of folders and feeds.
Previously we were composing the refresh URL with javascript, simply appending "/refresh" to the feed path. This does not work very well if the current path has parameters, for instance the locale; a feed path like "/feeds/1?locale=en" would give a refresh path like "/feeds/1?locale=en/refresh" which does not work at all (the request is routed to FeedsController#show instead of #refresh, and params[:locale]="en/refresh" which is a weird locale indeed.

I've opted for including the feed (or folder) refresh path in a data-* attribute in the feed/folder links in the sidebar. When the user clicks on a feed (or the "read all subscriptions" link under a folder), the data-refresh-path attribute of the link is copied to the data-refresh-path of the refresh button; this means that the path accessed via AJAX to refresh feeds and folders is originally composed in the server (in the _sidebar_feed and _sidebar_folder partials, to be specific).

To make sure the refresh URLs are correctly composed, I've changed the feed and folder routes to named routes (feed_refresh and folder_refresh, respectively). This enables the use of feed_refresh_path and folder_refresh_path helpers in the views, which always compose the correct paths, parameters and all.