Added the data-sidebar-feed attribute to all the links in the sidebar to read feeds and to read whole folders. The javascript that detects clicks on feeds in the sidebar now works by detecting the presence of this attribute. The value of the attribute is inconsequential, its presence is what matters to indicate that a link is a sidebar link to a feed.
Before this change, the javascript events detected the presence of the data-feed-id attribute. However this attribute was also present in some other elements (e.g. the Folder button) and this caused strange behaviors. It's much better to have an attribute that unequivocally indicates that a link is a sidebar link.