When successfully unsubscribing from a feed, the FeedsController returns an HTTP 204 code, instead of a 200. This is more in line with intended HTTP usage.
This will also be useful for the FoldersController: when removing a feed from a folder, a 204 will be returned if there are still feeds in the folder, and a 205 if there are no more feeds in the folder. This will tell the client whether to remove the folder from the view. This is not yet implemented.