WIP: user can remove a feed from a folder.
The way it will be implemented, the user can:

- click on "no folder" in the dropdown, removing feed from all folders
- if the feed is in a folder, click on it in the dropdown to remove the feed from this folder

Currently both behaviors are exactly the same, because currently feeds can belong to a single folder for a single user. However, by implementing it this way, it would be easier to change the app to allow feeds to belong to several folders, if we so decide in the future.

Currently implemented: routing, FoldersController#destroy method, skeleton of the Folder#remove_feed method, FoldersController unit tests.

Not yet implemented: most of the Folder#remove_feed method, unit tests for this method, all javascript necessary for this to work client-side, acceptance tests.