Implemented the Folder#remove_feed method. This methods removes a feed from a folder and, if no more feeds are in the folder, deletes it from the database.
In the end I've decided to remove support for passing the value "all" instead of a folder_id to the method. This would have meant passing the user id as an argument, because we never want to remove a feed from ALL folders, at most we could want to remove it from all folders belonging to a single user.

As this is a option that was not going to be used at the moment and which adds some unnecessary complexity, it's been removed from the roadmap for now.