Removing a feed from a folder is done with the FoldersController#remove method. The FoldersController#destroy method is unused and removed.
The new method is similar to the old one, with the difference that the folder id is not passed as a param, but instead is inferred (any feed can belong to at most one folder for the current user; this is enforced by methods in the Folder and Feed models).

The route is also changed; instead of the RESTful DELETE /folders/:id route, the DELETE /folders/remove route is used. It is mandatory that the feed_id is present in the params.

This change is desirable because of the way client interaction works. The client sends an order to remove a given feed from all folders, and it's not necessary that the client specifies the folder under which the feed is currently.