Fixed the alert that is shown when there is a problem removing a feed from folders. Also, fixed the read_feed coffeescript function.
The alert that was being shown was not the correct one.

The read_feed coffeescript function is responsible for selecting a feed in the sidebar and clicking on it, so the user reads its entries. The wrong CSS selector was being used, which resulted in the feed not being selected.