Fixed incorrect HTML in the sidebar.
The Start link is inside a <ul> that itself was nested inside a <ul>. This is not correct, a <ul> should contain <li> elements, not directly another <ul>. Browsers were not complaining, but I changed it anyway. Now both <ul>s are siblings instead of being nested one inside another.

Now the #sidebar id does not belong to the <ul> that contains folders, but instead to a <div> that encompasses both <ul>.

This has forced making changes in several acceptance tests, which were relying on CSS selectors which are no longer valid.