WIP: User can create a new folder and add the currently selected feed to it, by clicking on "New folder" in the dropdown.
Currently working: new folder is inserted in the sidebar.

Not yet working: insert folder in the dropdown, show alerts, remove empty folders.

An important new feature just implemented is that the FoldersController#create method returns JSON with embedded HTML. This is necessary because we need the POST create method that creates the new folder to return three values: the id of the new folder, the HTML to insert in the sidebar, and the HTML to insert in the dropdown. The only way to return from the server three values, two of which are HTML fragments, is to return JSON with the three values embedded in it. This has demanded some juggling with Rails, which is not intended for this use. If this technique gives satisfactory results I may use in other points of the application.