Fixed coffeescript code to handle the onClick event for the folders dropdown.
The problem was that the event handler was being associated with the dropdown elements when loading the page, and if new elements were dynamically added afterwards (by moving a feed to a new folder), they had no event handler associated and clicking on them did nothing.

The solution is to properly use the jQuery $.on() function to associate an event handler to a family of elements, even if they are dynamically added after loading the page.