Applied the change of commit 2475df52716adda6ada2d104b91b120c6669b88b to all coffeescript event handlers.
This means event handlers are defined in such a way that they are associated with their elements even if they are dynamically added after loading the page. This is not necessary in many cases right now, but it doesn't hurt and it could come handy if ever I decide to dynamically load some of those elements instead of being present in the page from the start.