Fix wrong event handler declaration in coffeescript, for the click event on the New Folder popup.
When applying the changes on dca2030ae423cc6070cb4001e7d5c4de132e44fd the line declaring this hander was half-changed between the old and the new format. The result was an invalid coffeescript line which resulted in invalid javacript and messed up the app's javascript code, making exceptions to be thrown and some handlers not work.