Added EntryState model.
This model represents the state in which a given feed entry is for a given user: read or unread.

Entries will be in the unread state when fetched from the feed server, and they will become read automatically when a user reads them. A user will also be able to mark entries as read or unread manually.

By default the view will show unread entries to the user, unless he manually indicates he wants to see all entries.