Added validation to ensure that a given entry has at most one state (read or unread) for a given user. Added callbacks to manage entry states when a user subscribes or unsubscribes from a feed.
When a user subscribes to a feed, EntryState instances are saved for each feed entry, to indicate that all those entries are unread for the user.

Conversely when a user unsubscribes from a feed, all EntryState instances for entries of that feed and belonging to that user are deleted; this means that the app does not store read/unread entries information for feeds from which the user has unsubscribed.