Removed the possibility of refreshing folders. From now on users can only refresh a single feed each time.
This simplifies code: the "refresh folder" was becoming a bit of a nightmare when implementing the "update unread entries count when refreshing" functionality. Refreshing folders is not a vital functionality, and so it's been declared out of scope right now.

Also, when refreshing a feed the entries list is no longer in the returned JSON. Instead of dynamically inserting the returned entries, the coffeescript event handler simulates a "click" event in the feed link, to force a reload of the entries through the normal FeedsController#show method. Again, this way code is simpler.