Implemented tests to detect regressions in a bug I'm currently experiencing: when a user clicks on a feed which is inside a folder, opens the Folder dropdown menu, and clicks on the same folder the feed already is at, an error message is shown. Worse, if it was the only feed in the folder, the folder is deleted and the feed left in the "no folder" category.
The tests (one unit test for the User model and one acceptance test for the "move feed to existing folder functionality) are currently failing, they are a first step towards fixing the bug and ensuring it doesn't appear again.

This bug is likely related to how empty folders are deleted by ActiveRecord callbacks as soon as they become empty. The solution is probably to control in the model or auxiliary class whether the old and new folders are the same when moving a feed, and in this case not doing anything.