Changed FolderFeedAdd.add_feed_to_folder so that when it's asked to add a feed to the folder it's already in, it doesn't actually make any changes in the database.
This fixes a bug in which the method removed the feed from the folder and then tried to add it back; but if the feed was the only one in the folder, thanks to the after_remove callback for the Folder.feeds association, the folder was deleted just after removing the feed from it. This meant that then FolderFeedAdd.add_feed_to_folder tried to add the feed to a folder that had just been deleted, which raised an error. From the point of view of the user the bug was quite confusing: he got an error message and then if he refreshed the page the folder had misteriously disappeared.

fixes #165