Fix the error that happened when reading a folder (or all subscribed feeds) and opening an entry.
The problem is that the data that was being sent to the server to mark the entry as read included a wrong feed id. Now unnecessary data is not sent from the client, and the return JSON contains arrays of changed feeds and folders, so that the sidebar links can be updated with the correct unread entry counts.

Also added a global function to mark as active a specific link in the sidebar. It is used so that the same link stays active after updating its unread entry count. It's also used when using the "mark all as read" button.

fixes #177