When opening an entry, the title link always opens in a new browser tab.
Changed this link: now an "open in new tab" icon is appended to the entry title inside the <a> tag, instead of having two <a> tags, one for the title and one for the icon.

Added CSS rule so that fontawesome icons do not get any text decoration. This is because when an icon was part of a link and the mouse hovered over the link, a broken underline appeared (the underline under the normal text was lower than the underline under the icon). With this change, the normal text in the link is underlined when hovered, but the icon is not. This looks much better.

Added acceptance test to ensure that entry title links are opened in new tabs.