A corner case when user tries to subscribe to a new feed is now controlled.
The corner case is the following:

- a feed is already in the database (the user may even be subscribed to it)
- a user submits the Add Subscription form with a URL that:
	a) leads to an HTML document, not to a feed
	b) is different from the URL saved in the DB for the already-existing feed (e.g. the feed in the database has url "http://www.meneame.net" but the user enters in the form "http://meneame.net", which actually serves the same HTML page)
	c) from the HTML the already-existing feed can be autodiscovered

Or to put it another way, the user tries to subscribe to a feed already in the database, entering a URL different than the URL that feed has in the database (but a URL that through feed autodiscovery leads to the already existing feed).

In this case an error message was shown in the view. Now this case is controlled and the user is subscribed to the already existing feed (or if he was already subscribed, a notice message is shown).