All links in entry summaries are opened in a new tab when the user clicks them.
To achieve this, a target="_blank" attribute is added to all such links before saving the entry in the database, replacing any previous target= attribute they could have.

Also if a user tries to subscribe to a feed he's already subscribed to, besides showing a notice message he's shown the Start page.