Prepared the app to send emails through Amazon SES.
The actual SMTP credentials for Amazon SES are not stored in Github, but rather are in a version of staging.rb manually copied to the deployment machine. On each deployment this version of staging.rb overwrites the one deployed from Github.

This means that from now on we will have to be careful to manually copy any changes staging.rb sufferts to the staging machine.