When deploying, capistrano links the correct environment configuration into the deployment.
Before this commit Capistrano always linked staging.rb into the deployment, even in the production environment, and it never linked production.rb in the production environment. This meant that the production environment tried to use the skeleton production config present in github, instead of the complete config that was copied by hand to the production server.