God gem is used to start and monitor the following background processes: Redis-server, resque worker for the :update_feeds queue, and resque scheduler.
When deploying with Capistrano, God is restarted in the remote server.

The God notification configuration is a per-environment file (it will be overwritten during deployment in staging and production with the version of the file that has been copied by hand to the server). This is because we intend to send email notifications through Amazon SES and I don't want SES SMTP credentials to be present in Github.