Pass the rails environment to the resque workers when starting them with God.
Redis configuration is not in the shared/redis deployment folder, but rather in the current/redis folder. This way it will be overwritten every time a deployment happens. However the aof and dump redis files are in the shared/redis folder, so that no redis data is lost when deploying.

Changed redis config file to better match the redis-server version installed in the production and staging servers.