Protected access to the Resque web console using the Cancan gem.
This has required adding an "admin" boolean attribute to the User model (and table). Only users that have the "admin" attribute with a true value will see the Resque link in the navbar, and any attempt to access its URL by non-admin users will be stopped.

The solution adopted is inspired in this blog post:

   http://simple10.com/resque-admin-in-rails-3-routes-with-cancan/