The perform_autodiscovery argument to the FeedClient.fetch method now defaults to false.
The only moment in which performing autodiscovery makes sense is when subscribing to a feed not yet present in the database. In the rest of points in which this method is invoked, autodiscovery should not be performed (it makes no sense trying to perform autodiscovery if the feed has been fetched in the past, because the correct fetch_url should be saved in the database).