Added support for Feedburner feeds that often don't include a GUID in their entries.
In this case the entry URL is used as a default value for its GUID (it's also used as default value for its title, a title is mandatory because we need to show it in the view). We need a GUID value for each entry so that when fetching we can detect if we're fetching a new entry or updating an already existing one.