Removed the unit tests that checked that the Entry model does not accept empty or nil values in the guid and title attributes.
Now that guid and title default to the same value as the url attribute when creating an Entry instance, these tests are no longer necessary. Actually the Entry model now DOES accept empty or nil values in those attributes, only those attributes are given a default value before saving in the database.