Added a new server with role "background" to the staging environment in Capistrano deployment script.
The idea is to use capistrano to deploy the app, precompile assets and restart Passenger only in the server with role :app, and to deploy the app and start god (which starts and monitors Redis and a Resque worker) only in the server with role :background. Also, secret files (database config and such) are only linked into the deployment in in those servers that make sense; e.g. there is no need to link the god notifications config in the :app server, because god will only run in the :background server.