Changed Resque configuration to account for the fact that we have two roles for servers:
- app servers will connect to the Redis instance in the background server
- background server will connect to its localhost Redis instance

The current architecture allows a single background server. If it becomes a bottleneck we will have to move to a different architecture; either with many background servers, each with their own Redis instance in some kind of master-server configuration, or with a single Redis server and many servers running Resque workers connected to the single Redis instance.