Changed Redis config so that it listens on all interfaces, not just in localhost.
This is a must now that resque workers and the web app are separated in different servers; jobs must be enqueued from the web server, while Redis is in the background server.

Security is enforced through EC2 security groups, so that the Redis port can only be accessed from the web server. It would be a huge security risk to open the Redis port to the world!

Perhaps in we should also use Redis authentication too...