Added FileClient class to save/delete files from the filesystem.
Changed the environment files so that either the S3Client or the FileClient class is assigned to the uploads_manager configuration variable. The production and Staging environments use the S3Client class, the rest use the FileClient class. The methods invoked are the same in both cases, so that for most of the code it is transparent whether uploaded files are stored in S3 or in local filesystem.