Add a before_save hook to the user model to ensure the encrypted_password field is encoded ad utf-8.
Hopefully this fixes the deluge of error lines that were appearing in Travis:

	[ERROR] Binary data inserted for `string` type on column `encrypted_password`

Inspired by https://github.com/padrino/padrino-framework/issues/1121 which leads to the line in rails where the error is logged and the encrypted_password is encoded as utf-8. Now we force the encoding beforehand, which should get rid of the error without affecting functionality at all.