Fixed acceptance test.
During acceptance testing, when an admin user tries to access the resque-web page, an HTTP 500 will be returned because Resque and Redis are not running during testing. However it's enough to check that a RoutingError is not raised to check that the app is allowing access to the user based on his admin role.