Fixed an error in tests that only manifested if tests were performed in a certain order.
This bug was probably present since commit 6c50d5f1b26ff29c112c7a83ed44c31a1cde4fb7

The problem was that the upload_manager object in the app configuration in some tests was being switched by a mock object. If later tests (for instance the acceptance test for the import OPML file feature) tried to use the upload manager after it had been switched by a mock object, tests fail because they try to use methods not defined in the mock (or the canned responses from the mock cause the tests to fail).

The solution is to never change the upload_manager object in the app config for another object (mock or otherwise). Instead some methods in the actual upload_manager object are stubbed as necessary during certain tests. Thanks to how stubbing scope works in Rspec, the stubbing is undone at the end of each test, and therefore subsequent tests are not affected.

For more about stubbing scope in Rspec see:

    https://www.relishapp.com/rspec/rspec-mocks/docs/scope