Added unit test for the SubscriptionsDataController. Also ensured that if any error is raised during importing, an ImportDataError is raised (instead of any other error class).
This ensures that the controller can react correctly and ensure that the user is always redirected to the start page instead of being left at a blank HTTP 500 error page. If any error is raised the data_import for the user will be left at status ERROR, and therefore the user will be shown a big red alert message telling him that something has gone wrong.