When adding a feed to a user's subscriptions, before marking each entry as unread for the user, ensure that the user does not already have an entry_state (read or unread) for that entry.
This is to guard against the possibility of a feed being added twice to a user's feeds. In this case, for those entries that already have an entry_state for the user, we don't have to add an entry_state record (in fact if we tried to add a second entry_state record for the same entry and user, a validation error would be raised).