When adding an entry to a feed's entries, before marking it as unread for each subscribed user, ensure that the user does not already have an entry_state (read or unread) for that entry.
This is to guard against the possibility of an entry being added twice to a feed's entries. In this case, for those users that already have an entry_state for the entry, we don't have to add an entry_state record (in fact if we tried to add a second entry_state record for the same entry and user, a validation error would be raised).

Related to commit 33acb6f433002aa7e30fd2dc5728bb1a621ec2e6. Fixes problems that were apparent since commit 6e5cdaa01b7bfeeb142e65a58edb9c173b5f3334.